Phase 21: Distributed Cosmology and Grafic IC Support 🌌🔭
Finalizing Phase 21 with full cosmological integration:
- Implemented Cosmology class with Friedman solver and growth factors.
- Added Grafic initial conditions reader for displacement and velocity loading.
- Integrated cosmological expansion (a, H) into Simulation loop and Courant condition.
- Hardened Poisson solver with comoving coordinate scaling and rho_tot subtraction.
- Fixed critical neighbor index propagation bug in AmrGrid.
- Improved memory safety in AMR tree traversal and particle management.

Benchmark verified on 128^3 grid with 2M particles. ✨🚀

+++ src/RamsesWriter.cpp
@@ -389,12 +389,12 @@ void RamsesWriter::write_particles(const AmrGrid& grid, const SnapshotInfo& info
         }
         {
             std::vector<int32_t> buf(npart);
-            for (int i = 0; i < npart; ++i) buf[i] = grid.idp[i];
+            for (int i = 0; i < npart; ++i) buf[i] = (int32_t)grid.idp[i];
             write_record_internal(file, buf.data(), npart);
         }
         {
             std::vector<int32_t> buf(npart);
-            for (int i = 0; i < npart; ++i) buf[i] = grid.levelp[i];
+            for (int i = 0; i < npart; ++i) buf[i] = (int32_t)grid.levelp[i];
             write_record_internal(file, buf.data(), npart);
         }
     }
